bump versions to 0.7.x (#12)
This updates the version of rustler used by this plugin to 0.37.x.
This is a minor (semver-breaking, since pre-1.0) release due as the
dependency we are upgrading also does not have this guarantee.

We update elixir multiple minor versions to ensure we keep the versions
of both packages in lockstep.